Backend/requirements.txt

google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.18.1
marshmallow==3.19.0
numpy==2.0.1
oauthlib==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyparsing==3.1.2
pytest==7.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.1.0
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
threadpoolctl==3.5.0
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.19
websockets==11.0.3
Werkzeug==2.3.4